Move typed span info into Sema
